requirements.txt
mcp
httpx
python-a2s
